refactor: move types to component folders
