+++ cloud-hub/src/components/Folder/types.ts
@@ -1,0 +1,10 @@
+export type Folder = {
+  id: string;
+  path: string;
+  files?: File[];
+  subFolders?: Folder[];
+  size: number;
+  numberOfFiles: number;
+  createdAt: Date;
+  updatedAt: Date;
+};
