+++ cloud-hub/src/components/User/types.ts
@@ -1,0 +1,9 @@
+export type User = {
+  id: string;
+  username: string;
+  email: string;
+  verified: boolean;
+  image?: string;
+  createdAt: Date;
+  updatedAt: Date;
+};
